+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Shared achievement links now show a preview of the achievement card with its geography and progress. (#3721)
+
 ## [1.15.2] - 2026-09-22, Berlin
 
 ### Added